Se  le corrigio el tipo de dato del identificador a cada repo , ya que en vez de int eran long. Y se crearon los endpoint de docentes, y preguntas

+++ src/main/java/uni/quindio/eduquizsolutions/repositories/AsignacionDiaRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import uni.quindio.eduquizsolutions.entities.AsignacionDias;
 @Repository
-public interface AsignacionDiaRepo extends JpaRepository<AsignacionDias, Integer> {
+public interface AsignacionDiaRepo extends JpaRepository<AsignacionDias, Long> {
 
 }
