+++ src/main/java/uni/quindio/eduquizsolutions/repositories/CursosRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Cursos;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface CursosRepo extends JpaRepository<Cursos, Integer> {
+public interface CursosRepo extends JpaRepository<Cursos, Long> {
 
 }
