+++ src/main/java/uni/quindio/eduquizsolutions/repositories/EdificioRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Edificios;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface EdificioRepo extends JpaRepository<Edificios, Integer> {
+public interface EdificioRepo extends JpaRepository<Edificios, Long> {
 
 }
