+++ src/main/java/uni/quindio/eduquizsolutions/repositories/EstadosRepo.java
@@ -1,10 +1,9 @@
 package uni.quindio.eduquizsolutions.repositories;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import uni.quindio.eduquizsolutions.entities.Estados;
 import org.springframework.stereotype.Repository;
+
 @Repository
-public interface EstadosRepo extends JpaRepository<Estados, Integer> {
+public interface EstadosRepo extends JpaRepository<Estados, Long> {
 
 }
