+++ src/main/java/uni/quindio/eduquizsolutions/repositories/EstudiantesRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Estudiantes;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface EstudiantesRepo extends JpaRepository<Estudiantes, Integer> {
+public interface EstudiantesRepo extends JpaRepository<Estudiantes, Long> {
 
 }
