+++ src/main/java/uni/quindio/eduquizsolutions/repositories/ExamenesRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Examenes;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface ExamenesRepo extends JpaRepository<Examenes, Integer> {
+public interface ExamenesRepo extends JpaRepository<Examenes, Long> {
 
 }
