+++ src/main/java/uni/quindio/eduquizsolutions/repositories/GruposRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Grupos;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface GruposRepo extends JpaRepository<Grupos, Integer> {
+public interface GruposRepo extends JpaRepository<Grupos, Long> {
    
 }
