+++ src/main/java/uni/quindio/eduquizsolutions/repositories/MateriaRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Materias;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface MateriaRepo extends JpaRepository<Materias, Integer> {
+public interface MateriaRepo extends JpaRepository<Materias, Long> {
 
 }
