+++ src/main/java/uni/quindio/eduquizsolutions/repositories/SalonesRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Salones;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface SalonesRepo extends JpaRepository<Salones, Integer> {
+public interface SalonesRepo extends JpaRepository<Salones, Long> {
 
 }
