+++ src/main/java/uni/quindio/eduquizsolutions/repositories/TemasExamene.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.TemasExamenes;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface TemasExamene  extends JpaRepository<TemasExamenes,Integer>{
+public interface TemasExamene  extends JpaRepository<TemasExamenes,Long>{
 
 }
